rfq-agent/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pydantic==2.6.4
pydantic-settings==2.2.1
python-multipart==0.0.9
aiofiles==23.2.1
PyMuPDF==1.24.1
openpyxl==3.1.2
python-dotenv==1.0.1
pillow==10.3.0
httpx==0.27.0
opencv-python-headless==4.10.0.84
anthropic==0.23.1
google-genai==1.47.0
ezdxf==1.4.2
